+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Changed
 
 * Replace dev.sh helper script with tox environments.
 * OAuth Grant Flows are now configured system-wide instead of per app.
+* Refactor OAuth2 application management views.
 
 `2.0`_ - 2021-10-29
 -------------------