+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Added
 ~~~~~
 
 * Provide an ``ExtensiblePolymorphicModel`` to support the features of extensible models for polymorphic models and vice-versa.
+* Add birthplace to Person model.
 
 Changed
 ~~~~~~~