+++ CHANGELOG.rst
@@ -23,6 +23,12 @@ Changed
 * OAuth Grant Flows are now configured system-wide instead of per app.
 * Refactor OAuth2 application management views.
 
+Credits
+~~~~~~~
+
+* We welcome new contributor 🐧 Jonathan Krüger!
+* We welcome new contributor 🐭 Lukas Weichelt!
+
 `2.0`_ - 2021-10-29
 -------------------