+++ CHANGELOG.rst
@@ -24,6 +24,11 @@ Changed
 * OAuth Grant Flows are now configured system-wide instead of per app.
 * Refactor OAuth2 application management views.
 
+Fixed
+~~~~~
+
+* Fix default admin contacts
+
 Credits
 ~~~~~~~